+++ sdk/ai/Azure.AI.Projects.Agents/tests/BetaVoiceAgentsConversationsTests.cs
@@ -5,9 +5,9 @@ using System.ClientModel;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Azure.AI.Projects.Agents._Beta.VoiceAgents;
 using Microsoft.ClientModel.TestFramework;
 using NUnit.Framework;
-using Azure.AI.Projects.Agents._Beta.VoiceAgents;
 
 #pragma warning disable AAIP001
 #pragma warning disable AAIP002